requirements/base.txt
argon2-cffi==19.2.0
Django==3.0.1
djangorestframework==3.11.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3